requirements.txt
aiogram
matrix-nio
httpx
fastapi
uvicorn
pydantic-settings
python-dotenv